+++ packages/onchainkit/src/minikit/index.ts
@@ -11,3 +11,5 @@ export { useClose } from './hooks/useClose';
 export { useComposeCast } from './hooks/useComposeCast';
 export { useViewCast } from './hooks/useViewCast';
 export { useIsInMiniApp } from './hooks/useIsInMiniApp';
+export { useSwapToken } from './hooks/useSwapToken';
+export { useSendToken } from './hooks/useSendToken';